Add sales equality comparator

+++ src/main/java/stat/domain/Product.java
@@ -58,12 +58,11 @@ public class Product {
         this.sales = sales;
     }
 
-
     @Override
     public boolean equals(Object obj) {
         if (obj instanceof Product) {
             Product product = (Product) obj;
-            return product.getProductId() == this.getProductId();
+            return product.getProductId().intValue() == this.getProductId().intValue();
         }
         return false;
     }
